Moved 'Serialization' and 'Exceptions' into Utils, references updated
