+++ src/interest/interest.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { SerializationModule } from '../serialization/serialization.module';
+import { SerializationModule } from '../utils/serialization/serialization.module';
 import { Interest } from './interest.entity';
 import { InterestController } from './interest.controller';
 import { InterestService } from './interest.service';
