+++ src/user/user.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { SerializationModule } from '../serialization/serialization.module';
+import { SerializationModule } from '../utils/serialization/serialization.module';
 import { UserController } from './user.controller';
 import { User } from './user.entity';
 import { UserService } from './user.service';
