+++ src/utils/exceptions/database.exception.ts
